Fix laravel.js delete inmueble

+++ resources/views/admin/inmueble/index.blade.php
@@ -17,12 +17,13 @@
                 </div>
             </div>
             <div class="bs-component">
-                <table class="table table-striped table-hover ">
+                <table class="table table-striped table-hover table-bordered">
                     <thead>
                         <tr>
                             <th>#</th>
                             <th>Tipo Propiedad</th>
                             <th>Descripcion</th>
+                            <th>Ciudad</th>
                             <th>Precio</th>
                             <th>Acciones</th>
                         </tr>
@@ -33,11 +34,13 @@
                                 <td>{{ $index + 1}}</td>
                                 <td>{{ $inmueble->nombre }}</td>
                                 <td>{{ $inmueble->descripcion }}</td>
+                                <td>{{ $inmueble->ciudad }}</td>
                                 <td>{{ $inmueble->precio}}</td>
                                 <td>
                                 	<a href="{{route('inmueble.show', $inmueble->id)}}" class="label label-primary">Detallar</a>
                                 	<a href="{{route('inmueble.edit', $inmueble->id)}}" class="label label-success">Editar</a>
-                                	<a href="{{route('inmueble.destroy', $inmueble->id)}}" class="label label-danger" data-method="delete">Eliminar</a>
+                                	<a href="{{route('inmueble.destroy', $inmueble->id)}}" class="label label-danger" data-method="delete" data-token="{{csrf_token()}}" data-confirm="Esta seguro de eliminar?">Eliminar</a>
+
                                 </td>
                             </tr>
                         @endforeach
